% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alueluokat_ja_kuntanumerot_2019}
\alias{Alueluokat_ja_kuntanumerot_2019}
\title{A data set containing area codes}
\format{
A data frame :
\describe{
\item{ Kuntanumero}{Kuntanumero}
\item{ Kunnannimi}{Kunnannimi}
\item{ Ruotsinkielilinennimi}{Ruotsinkielilinennimi}
\item{ Manner-SuomijaAhvenanmaa}{Manner-SuomijaAhvenanmaa}
\item{ NUTS2-koodi}{NUTS2-koodi}
\item{ NUTS2-alueennimisuuralue}{NUTS2-alueennimisuuralue}
\item{ AVI-nro}{AVI-nro}
\item{ AVI-nimi}{AVI-nimi}
\item{ Ely15nro}{Ely15nro}
\item{ Ely15-nimi}{Ely15-nimi}
\item{ Ely9nro}{Ely9nro}
\item{ Ely9-nimi}{Ely9-nimi}
\item{ Maakuntanro}{Maakuntanro}
\item{ Maakunnannimi}{Maakunnannimi}
\item{ Seutukuntanro}{Seutukuntanro}
\item{ Seutukunnannimi}{Seutukunnannimi}
\item{ Kuntamuotokoodi}{Kuntamuotokoodi}
\item{ Kuntamuoto}{Kuntamuoto}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
Alueluokat_ja_kuntanumerot_2019
}
\description{
A data set containing area codes
}
\keyword{datasets}
