% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erva2019}
\alias{erva2019}
\title{A data set containing map of Finland with ERVA (erityisvastuualue) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ erva_name}{ Name of erityisvastuualue district}
\item{ erva_code }{ Code of erityisvastuualue district}
\item{geom}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
erva2019
}
\description{
A data set containing map of Finland with ERVA (erityisvastuualue) boundaries
}
\keyword{datasets}
