% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyvinvointialueet2023}
\alias{hyvinvointialueet2023}
\title{Map of Finland with welfare counties (hyvinvointialueet) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ vuosi}{ Year }
\item{ hyvinvoint}{ Code }
\item{ nimi}{ Name in Finnish}
\item{ namn}{ Name in Swedish}
\item{ name }{ Name in English}
\item{geometry}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
hyvinvointialueet2023
}
\description{
A data set containing map of Finland with welfare counties (hyvinvointialueet) boundaries
}
\keyword{datasets}
