% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seutukunnat2019}
\alias{seutukunnat2019}
\title{Map of Finland with district (seutukunta) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ vuosi}{ Year }
\item{ seutukunta}{ Code }
\item{ nimi}{ Name in Finnish}
\item{ namn}{ Name in Swedish}
\item{ name }{ Name in English}
\item{geometry}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
seutukunnat2019
}
\description{
A data set containing map of Finland with district (seutukunta) boundaries
}
\keyword{datasets}
