% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHP2019}
\alias{SHP2019}
\title{A data set containing map of Finland with hospital district (sairaanhoitopiiri) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ sp_name}{ Name of hospital district}
\item{ sp_code }{ Code of hospital district}
\item{geom}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
SHP2019
}
\description{
A data set containing map of Finland with hospital district (sairaanhoitopiiri) boundaries
}
\keyword{datasets}
