% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suuralue2019}
\alias{suuralue2019}
\title{A data set containing map of Finland with regiona (suuralue) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ vuosi}{ Year }
\item{ suuralue}{ Code }
\item{ nimi}{ Name in Finnish}
\item{ namn}{ Name in Swedish}
\item{ name }{ Name in English}
\item{geometry}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
suuralue2019
}
\description{
A data set containing map of Finland with regiona (suuralue) boundaries
}
\keyword{datasets}
