% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaalipiiri2019}
\alias{vaalipiiri2019}
\title{A data set containing map of Finland with voting district (vaalipiiri) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ va_name}{ Name of voting district (vaalipiiri)}
\item{ va_code }{ Code of voting district (vaalipiiri)}
\item{geom}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
vaalipiiri2019
}
\description{
A data set containing map of Finland with voting district (vaalipiiri) boundaries
}
\keyword{datasets}
