% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mkMapFinland}
\alias{mkMapFinland}
\title{Make map of Finland}
\usage{
mkMapFinland(url)
}
\arguments{
\item{url}{code value (character)}
}
\value{
sf object
}
\description{
Make map of Finland
}
\note{
url must contain WSF Shapefile. Check available maps from http://geo.stat.fi/geoserver/web/
}
\examples{
hv.alueet<-mkMapFinland("https://tinyurl.com/hvalueet2022")
require(ggplot2)
ggplot(hv.alueet) +geom_sf()+ggtitle("Wellbeing services counties")
}
