% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suuralue2019}
\alias{suuralue2019}
\title{A data set containing map of Finland with regiona (suuralue) boundaries}
\format{
A data frame with map boundaries:
\describe{
\item{ vuosi}{ Year }
\item{ suuralue}{ Code }
\item{ nimi}{ Name in Finnish}
\item{ namn}{ Name in Swedish}
\item{ name }{ Name in English}
\item{  til_vuosi }{ til_vuosi}
\item{ vaesto  }{ vaesto}
\item{  vaesto_p }{vaesto_p }
\item{   miehet}{ miehet}
\item{  miehet_p }{ miehet_p}
\item{  naiset }{naiset }
\item{  naiset_p }{ naiset_p}
\item{  ika_0_14 }{ ika_0_14}
\item{   ika_0_14p}{ ika_0_14p}
\item{  ika_15_64 }{ ika_15_64}
\item{  ika_15_64p }{ ika_15_64p}
\item{  ika_65_ }{ ika_65_}
\item{   ika_65_p}{ ika_65_p}
\item{geometry}{Map polygon}
}
}
\source{
\url{http://geo.stat.fi/geoserver/}
}
\usage{
suuralue2019
}
\description{
A data set containing map of Finland with regiona (suuralue) boundaries
}
\keyword{datasets}
